+++ packages/app-admin-ui/server/env.js
@@ -1,8 +1,0 @@
-exports.mode = process.env.NODE_ENV === 'production' ? 'production' : 'development';
-
-exports.enableDevFeatures =
-  process.env.ENABLE_DEV_FEATURES === 'true'
-    ? true
-    : process.env.ENABLE_DEV_FEATURES === 'false'
-    ? false
-    : exports.mode !== 'production';