+++ packages/fields/src/types/Checkbox/views/Cell.js
@@ -1,7 +1,0 @@
-const CheckboxCellView = ({ data }) => {
-  if (data === true) return 'Checked';
-  if (data === false) return 'Unchecked';
-  return 'Not set';
-};
-
-export default CheckboxCellView;