+++ packages/fields/src/types/DateTimeUtc/index.js
@@ -1,4 +1,3 @@
-import DateTime from '../DateTime';
 import {
   DateTimeUtcImplementation,
   MongoDateTimeUtcInterface,
@@ -9,12 +8,6 @@ import {
 export default {
   type: 'DateTimeUtc',
   implementation: DateTimeUtcImplementation,
-  views: {
-    Controller: DateTime.views.Controller,
-    Field: DateTime.views.Field,
-    Filter: DateTime.views.Filter,
-    Cell: DateTime.views.Cell,
-  },
   adapters: {
     mongoose: MongoDateTimeUtcInterface,
     knex: KnexDateTimeUtcInterface,