+++ packages/fields/src/types/Virtual/views/Cell.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import PrettyData from '../prettyData';
-
-export default ({ data }) => {
-  return <PrettyData data={data} />;
-};