+++ packages/keystone/tests/Keystone.test.js
@@ -34,14 +34,10 @@ class MockFieldImplementation {
   getGqlAuxQueries() {
     return ['getFoo: Boolean'];
   }
-  extendAdminViews(views) {
-    return views;
-  }
 }
 
 const MockFieldType = {
   implementation: MockFieldImplementation,
-  views: {},
   adapters: { mock: MockFieldAdapter },
 };
 
@@ -67,7 +63,6 @@ describe('Keystone.createList()', () => {
   test('basic', () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);
 
@@ -92,7 +87,6 @@ describe('Keystone.createList()', () => {
   test('Reserved words', () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);
 
@@ -114,7 +108,6 @@ describe('Keystone.createList()', () => {
   test('plugins', () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);
 
@@ -159,7 +152,6 @@ describe('keystone.prepare()', () => {
   test('returns a Promise', () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);
 
@@ -171,7 +163,6 @@ describe('keystone.prepare()', () => {
   test('returns the middlewares array', async () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);
     const { middlewares } = await keystone.prepare();
@@ -182,7 +173,6 @@ describe('keystone.prepare()', () => {
   test('handles apps:undefined', async () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);
     const { middlewares } = await keystone.prepare({ apps: undefined });
@@ -193,7 +183,6 @@ describe('keystone.prepare()', () => {
   test('handles apps:[]', async () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);
     const { middlewares } = await keystone.prepare({ apps: [] });
@@ -204,7 +193,6 @@ describe('keystone.prepare()', () => {
   test('Handles apps without a `prepareMiddleware`', async () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);
     // For less-brittle tests, we grab the list of middlewares when prepare is
@@ -219,7 +207,6 @@ describe('keystone.prepare()', () => {
   test('filters out null middleware results', async () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);
     // For less-brittle tests, we grab the list of middlewares when prepare is
@@ -234,7 +221,6 @@ describe('keystone.prepare()', () => {
   test('filters out empty middleware arrays', async () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);
     // For less-brittle tests, we grab the list of middlewares when prepare is
@@ -249,7 +235,6 @@ describe('keystone.prepare()', () => {
   test('returns middlewares', async () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const middleware = jest.fn(() => {});
     const keystone = new Keystone(config);
@@ -264,7 +249,6 @@ describe('keystone.prepare()', () => {
   test('flattens deeply nested middlewares', async () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);
     const fn0 = jest.fn(() => {});
@@ -281,7 +265,6 @@ describe('keystone.prepare()', () => {
   test('should create `internal` GraphQL schema instance', async () => {
     const config = {
       adapter: new MockAdapter(),
-      cookieSecret: 'secretForTesting',
     };
     const keystone = new Keystone(config);